internal/pkg/rabbitmq/producer/send.go
package rabbitmq

import (
	"context"
	"encoding/json"
	"fmt"

	amqp "github.com/rabbitmq/amqp091-go"
)

// Send ...
func (p *Producer) Send(ctx context.Context, message interface{}) error {
	body, err := json.Marshal(message)
	if err != nil {
		return fmt.Errorf("JSON marshalling error: %w", err)
	}

	if err := p.rabbitMQChannel.PublishWithContext(ctx,
		"",
		p.queueName,
		false,
		false,
		amqp.Publishing{
			ContentType: "application/json",
			Body:        body,
		}); err != nil {
		return fmt.Errorf("message publish with context error: %w", err)
	}

	return nil
}
